Version 1.1

+++ src/main/java/com/desafiolecom/DTO/ServicoDto.java
@@ -11,7 +11,7 @@ public class ServicoDto implements Serializable{
 	
 	private String nome;
 	
-	private BigDecimal valor;
+	private double valor;
 
 	public Long getId() {
 		return id;
@@ -29,11 +29,11 @@ public class ServicoDto implements Serializable{
 		this.nome = nome;
 	}
 
-	public BigDecimal getValor() {
+	public double getValor() {
 		return valor;
 	}
 
-	public void setValor(BigDecimal valor) {
+	public void setValor(double valor) {
 		this.valor = valor;
 	}
